requirements.txt
Flask
python-dotenv
psycopg[binary]
flask-cors
gunicorn